backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.38
pydantic==2.10.6
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
email-validator==2.2.0
